Small refactor to reduce the time taken by tests and fix missing typings for embed functions on window
