+++ packages/embeds/embed-core/playwright/tests/embed-pages.e2e.ts
@@ -2,6 +2,8 @@ import { expect } from "@playwright/test";
 
 import { test } from "@calcom/web/playwright/lib/fixtures";
 
+import "../../src/types.d";
+
 test.describe("Embed Pages", () => {
   test("Event Type Page: should not have margin top on embed page", async ({ page }) => {
     await page.goto("http://localhost:3000/free/30min/embed");
@@ -45,111 +47,77 @@ test.describe("Embed Pages", () => {
     expect(marginFromTop).not.toBe(0);
   });
 
-  test("should change to embed when window.name is changed to cal-embed=", async ({ page }) => {
-    await page.goto("http://localhost:3000/free/30min");
+  test.describe("isEmbed, getEmbedNamespace, getEmbedTheme testing", () => {
+    test("when `window.name` is set to 'cal-embed=' and `theme` is supplied as a query param", async ({
+      page,
+    }) => {
+      const queryParamTheme = "dark";
+      await page.evaluate(() => {
+        window.name = "cal-embed=";
+      });
 
-    await page.evaluate(() => {
-      window.name = "cal-embed=";
+      await page.goto(`http://localhost:3000/free/30min?theme=${queryParamTheme}`);
+
+      const isEmbed = await page.evaluate(() => {
+        return window?.isEmbed?.();
+      });
+
+      const embedNamespace = await page.evaluate(() => {
+        return window?.getEmbedNamespace?.();
+      });
+
+      expect(embedNamespace).toBe("");
+      expect(isEmbed).toBe(true);
+      const embedTheme = await page.evaluate(() => {
+        return window?.getEmbedTheme?.();
+      });
+      expect(embedTheme).toBe(queryParamTheme);
+      const embedStoreTheme = await page.evaluate(() => {
+        return window.CalEmbed.embedStore.theme;
+      });
+      // Verify that the theme is set on embedStore.
+      expect(embedStoreTheme).toBe(queryParamTheme);
     });
 
-    await page.reload();
-
-    const isEmbed = await page.evaluate(() => {
-      return window?.isEmbed?.();
-    });
-    expect(isEmbed).toBe(true);
-  });
-
-  test("should return false on isEmbed when window.name does not contain cal-embed=", async ({ page }) => {
-    await page.goto("http://localhost:3000/free/30min");
-
-    await page.evaluate(() => {
-      window.name = "testing";
-    });
-
-    await page.reload();
-
-    const isEmbed = await page.evaluate(() => {
-      return window?.isEmbed?.();
-    });
-    expect(isEmbed).toBe(false);
-  });
-
-  test("should return 'testing' on getEmbedNamespace when window.name is changed to cal-embed=testing", async ({
-    page,
-  }) => {
-    await page.goto("http://localhost:3000/free/30min");
-
-    await page.evaluate(() => {
-      window.name = "cal-embed=testing";
+    test("when `window.name` does not contain `cal-embed=`", async ({ page }) => {
+      await page.evaluate(() => {
+        window.name = "testing";
+      });
+      await page.goto(`http://localhost:3000/free/30min`);
+      const isEmbed = await page.evaluate(() => {
+        return window?.isEmbed?.();
+      });
+      const embedNamespace = await page.evaluate(() => {
+        return window?.getEmbedNamespace?.();
+      });
+      expect(isEmbed).toBe(undefined);
+      expect(embedNamespace).toBe(null);
     });
 
-    await page.reload();
+    test("`getEmbedTheme` should use `window.CalEmbed.embedStore.theme` instead of `theme` query param if set", async ({
+      page,
+    }) => {
+      const theme = "dark";
+      await page.evaluate(() => {
+        window.name = "cal-embed=";
+      });
+      await page.goto("http://localhost:3000/free/30min?theme=dark");
+      let embedTheme = await page.evaluate(() => {
+        return window?.getEmbedTheme?.();
+      });
+      expect(embedTheme).toBe(theme);
 
-    const embedNamespace = await page.evaluate(() => {
-      return window?.getEmbedNamespace?.();
+      // Fake a scenario where theme query param is lost during navigation
+      await page.evaluate(() => {
+        history.pushState({}, "", "/free/30min");
+      });
+
+      embedTheme = await page.evaluate(() => {
+        return window?.getEmbedTheme?.();
+      });
+
+      // Theme should still remain same as it's read from `window.CalEmbed.embedStore.theme` which is updated by getEmbedTheme itself
+      expect(embedTheme).toBe(theme);
     });
-    expect(embedNamespace).toBe("testing");
-  });
-
-  test("should return empty string on getEmbedNamespace when window.name is changed to cal-embed=", async ({
-    page,
-  }) => {
-    await page.goto("http://localhost:3000/free/30min");
-
-    await page.evaluate(() => {
-      window.name = "cal-embed=";
-    });
-
-    await page.reload();
-
-    const embedNamespace = await page.evaluate(() => {
-      return window?.getEmbedNamespace?.();
-    });
-    expect(embedNamespace).toBe("");
-  });
-
-  test("should return window.CalEmbed.embedStore.theme on getEmbedTheme when window.name is changed to cal-embed=", async ({
-    page,
-  }) => {
-    const theme = "dark";
-
-    await page.goto("http://localhost:3000/free/30min");
-
-    await page.evaluate(() => {
-      window.name = "cal-embed=";
-    });
-
-    await page.reload();
-
-    await page.evaluate((theme) => {
-      window.CalEmbed.embedStore.theme = theme;
-    }, theme);
-
-    const embedTheme = await page.evaluate(() => {
-      return window?.getEmbedTheme?.();
-    });
-
-    expect(embedTheme).toBe(theme);
-  });
-
-  test("should return theme from query param on getEmbedTheme when window.name is changed to cal-embed= and window.CalEmbed.embedStore.theme is not available", async ({
-    page,
-  }) => {
-    const theme = "dark";
-
-    await page.goto(`http://localhost:3000/free/30min?theme=${theme}`);
-
-    await page.evaluate(() => {
-      window.name = "cal-embed=";
-    });
-
-    await page.reload();
-
-    const embedTheme = await page.evaluate(() => {
-      return window?.getEmbedTheme?.();
-    });
-
-    expect(embedTheme).toBe(theme);
   });
 });
